Fix RequestList deserialization from dict, add tests
